[WIP] Add credential dependency resolution.
Given an list of effective values (positive for input amounts, negative for
outputs), produce a graph of credential requests.

TODO
- [ ] consolidate reissuances of different request types (opportunistic parallel
      edges)

In principle this code should work for long lived credentials with minor tweaks
by introducing a 3rd credential type for round fees. Each request should be
allowed to convert input credits to fee credits, but not the other way around.
Although explicit in the credential dependencies, at the protocol level this
should be done by augmenting the balance proof with a range proof which keeps
the conversion amount hidden from the coordinator. Long term credentials are
then converted to epehemeral round tokens with the same flow as input
registration & confirmation, except that the converted amount should be hidden
so that the coordinator does not learn the value of the fee credentials used.